fix: added padding to user page
